sad module with only one file
